+++ nicovideo_api_client/api/v2/snapshot_search_api_v2.py
@@ -1,0 +1,11 @@
+from nicovideo_api_client.api.v2.targets import SnapshotSearchAPIV2TagsExact, SnapshotSearchAPIV2Keywords
+
+
+class SnapshotSearchAPIV2:
+    @staticmethod
+    def tags_exact() -> 'SnapshotSearchAPIV2TagsExact':
+        return SnapshotSearchAPIV2TagsExact()
+
+    @staticmethod
+    def keywords() -> 'SnapshotSearchAPIV2Keywords':
+        return SnapshotSearchAPIV2Keywords()
