+++ nicovideo_api_client/nvapi/v2/mylists.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from nicovideo_api_client.nicovideo.nvapi.v2.request import SnapshotSearchNVAPIV2Request
+from nicovideo_api_client.nvapi.v2.request import SnapshotSearchNVAPIV2Request
 
 
 class SnapshotSearchNVAPIV2Mylists:
