+++ nicovideo_api_client/nvapi/v3/limit.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from nicovideo_api_client.nicovideo.nvapi.v3.request import SnapshotSearchNVAPIV3Request
+from nicovideo_api_client.nvapi.v3.request import SnapshotSearchNVAPIV3Request
 
 
 class SnapshotSearchNVAPIV3Limit:
