+++ nicovideo_api_client/nvapi/v3/users.py
@@ -1,7 +1,7 @@
 from typing import Dict
 
-from nicovideo_api_client.nicovideo.constants import SortKeyType
-from nicovideo_api_client.nicovideo.nvapi.v3.sortkey import SnapshotSearchNVAPIV3SortKey
+from nicovideo_api_client.constants import SortKeyType
+from nicovideo_api_client.nvapi.v3.sortkey import SnapshotSearchNVAPIV3SortKey
 
 
 class SnapshotSearchNVAPIV3Users:
